Add ability to provide CLIP / CNAP information
And also add a stub in the UI for CDIP information, which is very
similar to CLIP / CNAP

The current behavior is for the modem to provide CLIP and CNAP if the
appropriate check box is checked.  If the box is checked but the line
edit is empty, then the number / name is assumed to be withheld.